Update for generate_share_transacrions

+++ sharechains/share.h
@@ -52,7 +52,7 @@ namespace c2pool::shares::share
         //===================================
 
         vector<uint256> new_transaction_hashes;
-        vector<tuple<unsigned long long, unsigned long long>> transaction_hash_refs; //TODO: check+test; # pairs of share_count, tx_count
+        vector<tuple<int, int>> transaction_hash_refs; //TODO: check+test; # pairs of share_count, tx_count
         uint256 far_share_hash;
         uint256 max_target; //from max_bits;
         uint256 target;     //from bits;
